updated the name of the crate in the TOML.
I had a different name for the project and
honestly I forgot about it and then I noticed the
old name was there. I believe there is another
Rust project using "Rusty nes" as it's name so I'm
updating the config files of the project to make
things more clear and consistent (and avoid
future problems).